Add files via upload
Added 2 new files. Product.java and ProductDetailsPage.java

Changes to CreatePages:
-	Added values to the hashmap that allow us to have more information on the details page
-	Removed Here from our search bar as it didn’t show the full word
-	Updated the itemDisplay function to now be able to use the photos that are uploaded as a button to get more information about the item.
o	To do this I added a cursor switch on line 151(to show it’s a button) and the actual action listener on 152. This listener on line 152 then runs a function that creates the ProductsDetialsPage if clicked
-	I had to change the way that the items were put onto the page because now more of the data including the file path are stored upon initialization

Aside from that the 2 new files are very straightforward in what they do. If you have any questions about it, please ask on discord